Fix race condition that ignores MAX_TRANSPORTS limit
The check for MAX_TRANSPORTS compares to the size of the list of
transports. Transports get added to that list after they have been
created. But creating a new Transport requires yielding, during which
some other callback chain may do run the comparison between
MAX_TRANSPORTS and the list of transports.

This happens for example when you kick off a bunch of requests to Riak,
all at the same time. Then each will request will pass through the
MAX_TRANSPORTS check, yield for a while and then append the new
Transport to the list. Each yield makes it possible for then next
request to pass the MAX_TRANSPORTS check, since the list is still
unmodified. The test in test_pbc_transport.py tests exactly this
scenario.

It would probably be a good idea to change the exception type that is
being raised to something more specific than Exception. How about
TooManyTransports?